requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysql==0.0.3
mysqlclient==2.2.7
pillow==12.1.0
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5

# Testing dependencies
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
Faker==33.4.0
pytest-factoryboy==2.7.0
